+++ core/revolve2/core/optimization/ea/generic_ea/__init__.py
@@ -8,9 +8,6 @@ from ._database import (
 )
 from ._optimizer import EAOptimizer
 
-#TMP!
-from ._optimizer_new import EAOptimizerNew
-
 __all__ = [
     "EAOptimizer",
     "DbEAOptimizer",
@@ -18,7 +15,5 @@ __all__ = [
     "DbEAOptimizerIndividual",
     "DbEAOptimizerParent",
     "DbEAOptimizerState",
-    "DbEnvconditions",
-
-    "EAOptimizerNew"
+    "DbEnvconditions"
 ]
