+++ experiments/default_study/bests_snap_2d.py
@@ -8,9 +8,9 @@ import numpy as np
 async def main() -> None:
 
     study = 'default_study'
-    experiments_name = ["plane", "tilted5", "tilted10"]
-    runs = list(range(1, 5+1))
-    generations = [110]
+    experiments_name = ["plane", "tilted5", "tilted10", "tilted15"]
+    runs = list(range(1, 10+1))
+    generations = [200]
     bests = 5
     path_out = f'/storage/karine/{study}/analysis/snapshots'
