+++ experiments/default_study/setup-experiments.sh
@@ -5,15 +5,14 @@
 
 study="default_study"
 # DO NOT use _
-experiments=("plane" "tilted5" "tilted10")
-seasons_conditions=("1.0_1.0_0" "1.0_1.0_5" "1.0_1.0_10")
-#experiments=("tilted15")
-#seasons_conditions=("1.0_1.0_15")
+
+experiments=("seasonal")
+seasons_conditions=("1.0_1.0_0#1.0_1.0_15")
 runs=10
 num_generations="200"
 
 # recommended 10-15
-num_terminals=15
+num_terminals=10
 
 mkdir data/${study}
