--- experiments/default_study_new/check-experiments.sh
@@ -1,37 +1,0 @@
-#!/bin/bash
-#set -e
-#set -x
-
-
-study="default_study"
-experiments=("plane" "tilted5" "tilted10")
-runs=10
-num_generations="200"
-
-
-# discover unfinished experiments
-
-to_do=()
-for i in $(seq $runs)
-do
-    run=$(($i))
-
-    for experiment in "${experiments[@]}"
-    do
-
-     printf  "\n${experiment}_${run} \n"
-     file="/storage/karine/${study}/${experiment}_${run}.log";
-
-     #check experiments status
-     if [[ -f "$file" ]]; then
-
-            lastgen=$(grep -c "Finished generation" $file);
-            echo " latest finished gen ${lastgen}";
-
-     else
-         # not started yet
-         echo " None";
-     fi
-
-    done
-done
