--- experiments/default_study_new/run-experiments.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-screen -d -m -S run_loop ./experiments/default_study/setup-experiments.sh
